refactor(tools): move the shared parity helpers to a neutral module
